% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/jn.R
\name{jn}
\alias{floodlight}
\alias{jn}
\title{Johnson-Neyman Technique}
\usage{
jn(model, dv, iv, mod, mrange, alpha = 0.05, yas = "none")
}
\arguments{
\item{model}{Regression model (lm, glm, list).}

\item{dv}{Dependent variable (character).}

\item{iv}{Independent variable (character).}

\item{mod}{Moderator variable(s) (character or character vector).}

\item{mrange}{Range of values that jn should examine for moderator variable. Uses the current range of moderator values by default (numeric vector).}

\item{alpha}{Alpha level to use (numeric).}

\item{yas}{Show y (or conditional effect) as: \code{"none", "ratio","probability","percentage"}, \code{yas="none"} by default.}
}
\value{
A list with the elements
}
\description{
Probe moderation effect using the Johnson-Neyman technique
}
\examples{
\dontrun{
myModel <- lm('DV ~ IV + MOD', data=someData)
jnresults <- jn(myModel, dv='DV', iv='IV', mod='MOD')
jnresults <- jn(myModel, dv='DV', iv='IV', mod='MOD', alpha=.01)
plot(jnresults)
}
}
\references{
Spiller, S. A., Fitzsimons, G. J., Lynch, J. G., Jr, & McClelland, G. H. (2013). Spotlights, floodlights, and the magic number zero: Simple effects tests in moderated regression. Journal of Marketing Research, 50(2), 277-288.

Bauer, D. J., & Curran, P. J. (2005). Probing interactions in fixed and multilevel regression: Inferential and graphical techniques. Multivariate Behavioral Research, 40(3), 373-400.
}

