% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/print.jn.R
\name{print.jn}
\alias{print.jn}
\title{Print Function For Johnson-Neyman}
\usage{
\method{print}{jn}(x, \dots)
}
\arguments{
\item{x}{An object of class \code{"jn"}.}

\item{\dots}{Additional arguments (not supported yet).}
}
\value{
none
}
\description{
Prints function for objects of class \code{"jn"}
}
\examples{
\dontrun{
myModel <- lm('DV ~ IV + MOD', data=someData)
jnresults <- jn(myModel, dv='DV', iv='IV', mod='MOD')
jnresults
}
}

