% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/print.pickapoint.R
\name{print.pickapoint}
\alias{print.pickapoint}
\title{Print Function For Pick-A-Point}
\usage{
\method{print}{pickapoint}(x, \dots)
}
\arguments{
\item{x}{An object of class \code{"pickapoint"}.}

\item{\dots}{Additional arguments (not supported yet).}
}
\value{
none
}
\description{
Print function for objects of class \code{"pickapoint"}
}
\examples{
\dontrun{
myModel <- lm('dv ~ iv + mod', data=someData)
papresults <- pickapoint(myModel, dv='DV', iv='IV', mod='MOD')
papresults
}
}

