% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plot.jn.R
\name{plot.jn}
\alias{plot.jn}
\title{Plot Function For Johnson-Neyman}
\usage{
\method{plot}{jn}(x, xlab = "", ylab = "", xlim = 0, ylim = 0, \dots)
}
\arguments{
\item{x}{an object of class \code{"jn"}.}

\item{xlab}{a title for the x axis (character).}

\item{ylab}{a title for the y axis (character).}

\item{xlim}{coordinates range for x axis (numeric vector).}

\item{ylim}{coordinates range for y axis (numeric vector)}

\item{\dots}{Additional arguments (not supported yet).}
}
\value{
none
}
\description{
Plot function for objects of class \code{"jn"}.
}
\examples{
\dontrun{
myModel <- lm('DV ~ IV + MOD', data=someData)
jnresults <- jn(myModel, dv='DV', iv='IV', mod='MOD')
plot(jnresults)
}
}

